feat: handle list of filenames in `absolute_path` and `file_mtime` (#89)
Closes #88

### Summary of Changes

`file_mtime` and `absolute_path` can now work with a list of strings as
well as a single string. This is needed for function like
`ImageList.from_files` that accept multiple paths.

For now, order matters. Specifying the same paths in a different order
leads to a different memoization key.