feat: prepare and pool processes (#87)
Closes #85

### Summary of Changes

- Use a process pool to keep started processes waiting
- The max. amount of pipeline processes is now set to `4`.
- Reuse started processes. This should be correct, as the same pipeline
process cannot be used by multiple pipelines at the same time. As the
`metapath` is reset to remove the custom generated Safe-DS pipeline
code, only global library imports (and settings) should remain. If this
is a concern, `maxtasksperchild` can be set to `1`, in which case
pipeline processes are not reused.
- Reuse shared memory location for saving placeholders, if the
memoization infrastructure has added such a location to the object being
saved

---------

Co-authored-by: megalinter-bot <129584137+megalinter-bot@users.noreply.github.com>
Co-authored-by: Lars Reimann <mail@larsreimann.com>