fix: encoding of non-finite floats (#190)
### Summary of Changes

By default, the `dumps` method of Python's `json` module generates
invalid JSON for non-finite floats (NaN, +/- Infinity). We now convert
them to a string instead. For tables, we still replace them with nulls
as before to avoid mixed data.

If we later use a binary format to transfer the table data, we won't
need to replace these values in tables anymore.